back/requirements.txt
Django==4.2
django-two-factor-auth==1.17.0
django-otp==1.5.4
django-formtools==2.3
django-phonenumber-field==6.4.0
qrcode==7.4.2
psycopg2-binary>=2.8
djangorestframework
environs